Enable access to files by title

+++ server/fileCache.mts
@@ -1,4 +1,5 @@
 import { basename } from "node:path"
+import { parse as parseHtml } from "node-html-parser"
 import { applyTemplating, type Meta } from "./dom.mts"
 import {
   createFileAndDirectories,
@@ -7,37 +8,40 @@ import {
   type MyDirectoryEntry,
   readFile,
   readFileRaw,
+  type ReadResults,
   removeFile,
   updateFile,
 } from "./filesystem.mts"
 import debug from "debug"
+import { MissingFileQueryError } from "./error.mts"
+import { renderMarkdown } from "./utilities.mts"
 const log = debug("server:fileCache")
 
 export type FileContentsAndDetails = {
   meta: Meta
-  originalContent: { content: string; foundInDirectory: string }
+  originalContent: { content: string; foundInDirectory: string; buffer: Buffer }
+  renderability: Renderability
 } & MyDirectoryEntry
+
 export type FileCache = {
   listOfFilesAndDetails: Array<FileContentsAndDetails>
   getByContentPath: (path: string) => FileContentsAndDetails | undefined
   getByTitle: (title: string) => FileContentsAndDetails | undefined
+  getByContentPathOrContentTitle: (
+    pathOrTitle: string,
+  ) => FileContentsAndDetails | undefined
   readFile: (path: string) => ReturnType<typeof readFile>
   readFileRaw: (path: string) => ReturnType<typeof readFileRaw>
   fileExists: (path: string) => ReturnType<typeof fileExists>
   createFileAndDirectories: (params: {
-    directory: string
     contentPath: string
     content: string
   }) => ReturnType<typeof createFileAndDirectories>
   updateFile: (params: {
-    directory: string
     contentPath: string
     content: string
   }) => ReturnType<typeof updateFile>
-  removeFile: (params: {
-    directory: string
-    contentPath: string
-  }) => ReturnType<typeof removeFile>
+  removeFile: (params: { contentPath: string }) => ReturnType<typeof removeFile>
 }
 
 export const buildEmptyCache = async (): ReturnType<typeof buildCache> => {
@@ -45,6 +49,7 @@ export const buildEmptyCache = async (): ReturnType<typeof buildCache> => {
     listOfFilesAndDetails: [],
     getByTitle: () => undefined,
     getByContentPath: () => undefined,
+    getByContentPathOrContentTitle: () => undefined,
     readFile: () => {
       throw new Error("No files exist in empty cache")
     },
@@ -69,7 +74,49 @@ export const buildCache = async ({
 }: {
   searchDirectories: string[]
 }): Promise<FileCache> => {
-  const listOfFilesAndDetails = await getContentsAndMetaOfAllFiles({
+  if (searchDirectories.length === 0) {
+    throw new Error("Cache requires non-empty searchDirectories upfront")
+  }
+  let listOfFilesAndDetails: FileContentsAndDetails[] = []
+  const filesByTitle: Record<string, FileContentsAndDetails> = {}
+  const filesByContentPath: Record<string, FileContentsAndDetails> = {}
+  const addFileToCacheData = (everything: FileContentsAndDetails) => {
+    listOfFilesAndDetails.push(everything)
+    filesByContentPath[everything.contentPath] = everything
+
+    if (typeof everything.meta.title === "string") {
+      filesByTitle[everything.meta.title] = everything
+    } else if (everything.meta.title) {
+      log(`Title must be a string, got %o`, everything.meta.title)
+      throw new Error(`Title must be a string, see log`)
+    }
+  }
+  const removeFileFromCacheData = async (contentPath: string) => {
+    listOfFilesAndDetails = listOfFilesAndDetails.filter(
+      ({ contentPath }) => contentPath !== contentPath,
+    )
+    const detail = filesByContentPath[contentPath]
+    delete filesByContentPath[contentPath]
+    if (typeof detail.meta.title === "string")
+      delete filesByTitle[detail.meta.title]
+
+    // Find if there's a revealed shadow file
+    const existsResults = await fileExists({ contentPath, searchDirectories })
+    if (existsResults.exists) {
+      const templateResults = await getFileContentsAndMetadata({
+        contentPath,
+        searchDirectories,
+        fileCache,
+      })
+      addFileToCacheData({
+        name: basename(contentPath),
+        type: "file",
+        contentPath,
+        ...templateResults,
+      })
+    }
+  }
+  const allFiles = await getContentsAndMetaOfAllFiles({
     // TODO: To recover from race conditions on initial build,
     // in the future, probably want to be able to start with the last cache.
     // Except that wouldn't account for deletions? Unless that was repaired first?
@@ -77,38 +124,39 @@ export const buildCache = async ({
     searchDirectories,
   })
 
-  const filesByContentPath: Record<string, FileContentsAndDetails> = {}
-  listOfFilesAndDetails.forEach((everything) => {
-    filesByContentPath[everything.contentPath] = everything
-  })
-  const filesByTitle: Record<string, FileContentsAndDetails> = {}
-  listOfFilesAndDetails.forEach((everything) => {
-    if (typeof everything.meta.title !== "string") return
-    filesByTitle[everything.meta.title] = everything
-  })
+  allFiles.forEach(addFileToCacheData)
+
   const fileCache: FileCache = {
     listOfFilesAndDetails,
     getByContentPath: (path) => filesByContentPath[path],
     getByTitle: (title) => filesByTitle[title],
-    readFile: (path) =>
-      // TODO: Read the file from the cache
-      readFile({
-        searchDirectories,
-        contentPath: path,
-      }),
-    readFileRaw: (path) =>
-      // TODO: Read from the cache
-      readFileRaw({
-        searchDirectories,
-        contentPath: path,
-      }),
+    getByContentPathOrContentTitle: (pathOrTitle) => {
+      return pathOrTitle === "/"
+        ? filesByContentPath["/index.html"]
+        : (filesByTitle[decodeURIComponent(pathOrTitle).replace(/^\//, "")] ??
+            filesByContentPath[pathOrTitle])
+    },
+    readFile: async (path) => {
+      const entry = filesByContentPath[path]
+      if (!entry) {
+        throw new MissingFileQueryError(path)
+      }
+      return entry.originalContent
+    },
+    readFileRaw: async (path) => {
+      const entry = filesByContentPath[path]
+      if (!entry) {
+        throw new MissingFileQueryError(path)
+      }
+      return entry.originalContent
+    },
     fileExists: async (path) =>
       filesByContentPath[path]
         ? { exists: true, ...filesByContentPath[path].originalContent }
         : { exists: false },
-    createFileAndDirectories: async ({ directory, contentPath, content }) => {
+    createFileAndDirectories: async ({ contentPath, content }) => {
       const result = await createFileAndDirectories({
-        directory,
+        directory: searchDirectories.at(0)!,
         contentPath,
         content,
       })
@@ -128,34 +176,54 @@ export const buildCache = async ({
       // deeper in the  stack of shadows, and maybe there's still a file higher
       // up that should shadow it. So maybe we shouldn't always be adding it
       // to these structures
-      listOfFilesAndDetails.push(details)
-      filesByContentPath[contentPath] = details
-      if (details.meta.title) {
-        if (typeof details.meta.title !== "string") {
-          log(`Title must be a string, got %o`, details.meta.title)
-          throw new Error(`Title must be a string, see log`)
-        }
-        filesByTitle[details.meta.title] = details
-      }
+      addFileToCacheData(details)
 
       return result
     },
 
-    // TODO: Update cache. The "original contents" need to change at least
-    updateFile: async ({ directory, contentPath, content }) =>
-      updateFile({ directory, contentPath, content }),
-
-    removeFile: async ({ directory, contentPath }) => {
-      // TODO: Update the cache. This might mean searching for an uncovered
-      // previously-shadowed core file. Maybe search the searchDirectories again
-      // and start from there..
-      return removeFile({ directory, contentPath })
+    updateFile: async ({ contentPath, content }) => {
+      const directory = searchDirectories.at(0)!
+      const result = await updateFile({
+        directory,
+        contentPath,
+        content,
+      })
+      await removeFileFromCacheData(contentPath)
+      const templateResults = await getFileContentsAndMetadata({
+        contentPath,
+        searchDirectories,
+        fileCache,
+      })
+      addFileToCacheData({
+        name: basename(contentPath),
+        type: "file",
+        contentPath,
+        ...templateResults,
+      })
+      return result
+    },
+    removeFile: async ({ contentPath }) => {
+      const directory = searchDirectories.at(0)!
+      const existingEntry = filesByContentPath[contentPath]
+      if (!existingEntry) {
+        throw new MissingFileQueryError(contentPath)
+      }
+      if (existingEntry.originalContent.foundInDirectory !== directory) {
+        throw new Error(
+          "Can only delete files in the top-level search directory",
+        )
+      }
+      // Don't update the cache until the operation is successful
+      const result = await removeFile({ directory, contentPath })
+      await removeFileFromCacheData(contentPath)
+      return result
     },
   }
 
   return fileCache
 }
 
+export type Renderability = "html" | "markdown" | "static"
 const getFileContentsAndMetadata = async ({
   contentPath,
   searchDirectories,
@@ -164,7 +232,11 @@ const getFileContentsAndMetadata = async ({
   fileCache: FileCache
   contentPath: string
   searchDirectories: string[]
-}) => {
+}): Promise<{
+  originalContent: ReadResults
+  meta: Meta
+  renderability: Renderability
+}> => {
   const readResults = await readFile({
     searchDirectories,
     contentPath,
@@ -181,6 +253,7 @@ const getFileContentsAndMetadata = async ({
       return {
         ...result,
         originalContent: readResults,
+        renderability: "html",
       }
     } catch (error) {
       throw new Error(
@@ -188,10 +261,26 @@ const getFileContentsAndMetadata = async ({
       )
     }
   } else if (/\.md$/.test(contentPath)) {
-    // TODO: Parse and separate frontmatter of markdown file
-    return { originalContent: readResults }
+    try {
+      const meta: Meta = {}
+      const markdownContent = renderMarkdown(readResults.content)
+      const root = parseHtml(markdownContent)
+      const h1 = root.querySelector("h1")
+      if (h1) {
+        meta.title = h1.innerText
+      }
+      return {
+        meta,
+        originalContent: readResults,
+        renderability: "markdown",
+      }
+    } catch (error) {
+      throw new Error(
+        `Couldn't apply templating for '${contentPath}': ${error}`,
+      )
+    }
   } else {
-    return { originalContent: readResults }
+    return { originalContent: readResults, meta: {}, renderability: "static" }
   }
 }
 
@@ -234,8 +323,7 @@ export const resolveDirEntToAllStuff = async ({
   })
   return {
     ...dirent,
-    type: "file", // TypeScript doesnt track the filter above
-    meta: "meta" in templateResults ? templateResults.meta : {},
-    originalContent: templateResults.originalContent,
+    type: "file",
+    ...templateResults,
   } as const
 }
